Add files via upload
Watch videos from 1 to 3  from the following link. Download the new resources via this link
We only care about backend , so understand the frontend with a pinch of salt. Watch video 4 to 6 and make the products details page like the udemy trainer.
Make the details page by watching video 7. 
Only after doing the previous steps, Watch this video to understand Dynamc routes in depth and the truth about skipping EJS