Add files via upload
Add a delete button. (This is very similar to edit functionality)
On click of the delete button call a new route ==> /delete/{id of product}
Define this new route in your backend. Take the id.
Now simply add a function in Product model called deleteproductbyID. Simply read from the file and remove the element
Now reload the product page you shouldnot be able to see the product. Awesome